Fix the click to go to line in textbox
This started as a fix to fix it so you could mark the info in the
textbox and copy it out. To fix this i removed word wrap so line count
matches no matter the size and then implemented that if the row does not
start with row then it will not send you to the datagrid and change the
focus.