Cleaning the way we handle API responses.
Adding support for BusinessRuleApiResponse code 100. This code will flag
the line as failed and move on like with all other business rule errors.
Changed from if else to switch statements for a little performance and
to make easier to read.
Also changed it so the method does not take the current error count just
returns number of errors and they are then added to the total where the
method is called.