Net core 9 (#9)
* update to use .net 9

The dataimporter is currently running .netcore3.1 that is a very old
version. Moving the project to .net 9 so we are on the latest version

* Adding caching

The dataimporter can be very slow to start because it fetches a lot of
data for each tab. To make it a bit faster i have added caching so it
only loads each type from the API on startup once.

* update to use .net 9

The dataimporter is currently running .netcore3.1 that is a very old
version. Moving the project to .net 9 so we are on the latest version

* Adding caching

The dataimporter can be very slow to start because it fetches a lot of
data for each tab. To make it a bit faster i have added caching so it
only loads each type from the API on startup once.

* Update nuget packages

The project had some outdated packages that needed to be updated.
This commit changes the IdentityModel packages to the new version.
And updates the WebUtilities to latest version.

* Remove the login using loginportal

Change it so the DataImporter only uses personal access tokens.

* Fix the click to go to line in textbox

This started as a fix to fix it so you could mark the info in the
textbox and copy it out. To fix this i removed word wrap so line count
matches no matter the size and then implemented that if the row does not
start with row then it will not send you to the datagrid and change the
focus.

* Remove unused Duende code

After removing the option to login using the login portal the Duende
nuget package was no longer needed, so removing the code.

* Cleaning the way we handle API responses.

Adding support for BusinessRuleApiResponse code 100. This code will flag
the line as failed and move on like with all other business rule errors.
Changed from if else to switch statements for a little performance and
to make easier to read.
Also changed it so the method does not take the current error count just
returns number of errors and they are then added to the total where the
method is called.

* Add submit when you press enter in the PAT textbox

Added is so when you press enter in the PAT textbox it submits the form.
This way you don't need to press the login using PAT button.

* Fixing the design.

There where some overlapping textboxes and some labels that went over
the textbox next to them. This commit fixes that so it looks a better.

* Add link to PAT site.

There where some concerns that after removing the login via LoginPortal
people would have a hard time finding the site to create personal access
tokens so added a link to the personal access token page.

* Cleaning up the cache

Change it to be an absolute timeout and not a sliding timeout.
Added it so that when logout it clears the cache.

* fix using

Fix it so the System.Reflection is not using but part of the method.

* Fix the login transition.

When logging in the login form disappeared and the main form didn't show
up until all the initial data was loaded. To fix this i am changing the
text on the login button and disabling it so you can see that something
is happening while the system is working.

* Layout clean up.

There where some layout elements that where a bit to big so have gone
through all the tabs and cleaned them up a bit.

* Fix spelling error in help text

Laura found a small spelling error in the help text. This has now been
fixed.

* Sort switch cases

Sorting switch cases so they are in increasing order.