fix issue with delimiter dropdown (#10)
All our delimiter comboboxes where set to the DropDown style, this
allows the user to manually input what ever they want and not only
select the allowed values. If they do this it breaks when they try to
select a file since the delimiter is given as a input.
To fix this the style has been changed to DropDownList.