Fix hard error when importing time registrations (#11)
* Fix hard error when importing time registrations

When importing time registrations you could encounter an issue where you
got an error that caused the validation to stop. This was caused by a
null pointer exception. I have added some checks for null.

* Tasks should be cached per project

Tasks are fetched per project and should therefor also be cached per
project. This commit adds the ProjectId to the cache key.

* Better error handling

We had an issue where Entities could be null when getting a response
back from the API. To ensure that the DataImporter can still open, I
have added exception handling that will inform the user that there is an
error but will then continue.

* Fix comments from Reviewed-by:

ensure formatting it the same and use the same way of doing null checks.

* Fix more formatting.

fix more formatting of else and catch statements.